1、添加了TableInfoData的Swagger注解 2、完善了各种容错校验，学科的列表查询校验 3、添加了对时段试卷的结束时间和开始时间的校验

+++ ms_stu_pro_320_server_new/ruoyi-manage/src/main/java/com/mashang/controller/TestRecordController.java
@@ -30,7 +30,7 @@ public class TestRecordController extends BaseController {
 
     @GetMapping("/page")
     @ApiOperation("根据提交时间倒叙分页查询该学生所有做过的试卷基本信息")
-    public TableDataInfo<List<TestRecordListVo>> pageTestRecord(@Validated PageQuery pageQuery, Integer subjectId){
+    public TableDataInfo<List<TestRecordListVo>> pageTestRecord(@Validated PageQuery pageQuery, Long subjectId){
         TestRecordQuery testRecordQuery = new TestRecordQuery();
         Long userId = SecurityUtils.getUserId();
         testRecordQuery.setUserId(userId);
