+++ ms_stu_pro_320_server_new/ruoyi-manage/src/main/java/com/mashang/domain/query/student/TestPageQuery.java
@@ -10,6 +10,6 @@ public class TestPageQuery {
     private Integer testType;
 
     @ApiModelProperty(value = "学科id")
-    private Integer subjectId;
+    private Long subjectId;
 
 }
